[Chore] NavigationContext 분리
